client/serializers.py
from rest_framework import serializers
from .models import Address, Client

# Serializers define the API representation.

class AddressSerializer(serializers.ModelSerializer):

    class Meta:
        model = Address
        fields = "__all__"

class ClientSerializer(serializers.ModelSerializer):

    class Meta:
        model = Client
        fields = "__all__"
